refactor(event): move decoding helpers into SDK parser
