+++ sdk-libs/event/src/lib.rs
@@ -16,7 +16,10 @@ pub mod indexed;
 pub mod instruction;
 pub mod reconstruct;
 
-pub use deposit::{proofless_output, proofless_outputs};
+pub use deposit::{
+    decode_encrypted_ring_deposit_output_data, decode_output_data, proofless_output,
+    proofless_outputs,
+};
 pub use indexed::{
     event_kind_from_indexed, general_event_from_indexed, indexed_events_from_instruction_groups,
     instruction_may_emit_events, IndexedEvent,
@@ -26,4 +29,33 @@ pub use reconstruct::{
     merge_general_event, reconstruct_general_event, reconstruct_general_event_from_payload,
     transact_general_event,
 };
-pub use zolana_event::EventDecodeError;
+
+/// Why an emitted event or output payload could not be decoded or rebuilt.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum EventDecodeError {
+    MissingInstructionTag,
+    InvalidInstructionTag(u8),
+    InvalidPayload,
+    InvalidEventKind(u8),
+    InvalidOutputData,
+    MissingOutput,
+    MissingDepositSplTransfer,
+    /// The emitting instruction's data did not parse as the instruction the
+    /// event kind belongs to.
+    InvalidSourceInstructionData,
+    /// The emitting instruction's tag cannot produce this event kind.
+    UnsupportedSourceInstruction(u8),
+    /// An output's `OwnerTag::Account` index points past the emitting
+    /// instruction's account list.
+    OutputOwnerAccountMissing(u8),
+    /// The event names more input trees than the instruction data can assign
+    /// inputs to.
+    UnsupportedInputTreeCount(usize),
+    /// `spl_transfers` and the instruction's interface transfers differ in length.
+    SplTransferCountMismatch,
+    /// The event kind carries no [`GeneralEvent`](zolana_event::GeneralEvent) view
+    /// (nullifier-tree updates).
+    NotAGeneralEvent,
+    /// Queue sequence or leaf index arithmetic overflowed.
+    IndexOverflow,
+}
