+++ sdk-libs/event/tests/reconstruct.rs
@@ -8,12 +8,12 @@ use support::{
     source, transact_ix, transact_source, INPUT_TREE, OUTPUT_TREE, SALT, TX_VIEWING_PK,
 };
 use zolana_event::{
-    tag, EventDecodeError, EventKind, GeneralEvent, Input, InputTreeSequence, MessageData,
-    NullifierTreeUpdateEvent, OutputUtxo, SplTransfer, TransactEvent,
+    tag, EventKind, GeneralEvent, Input, InputTreeSequence, MessageData, NullifierTreeUpdateEvent,
+    OutputUtxo, SplTransfer, TransactEvent,
 };
 use zolana_event_parser::{
-    indexed_events_from_instruction_groups, reconstruct_general_event, InstructionGroup,
-    ParsedInstruction,
+    indexed_events_from_instruction_groups, reconstruct_general_event, EventDecodeError,
+    InstructionGroup, ParsedInstruction,
 };
 use zolana_interface::instruction::{InterfaceTransfer, OwnerTag, TransactOutput};
